# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DFD API",
        "contact": {
            "name": "Seokho Song",
            "url": "http://github.com/devsdkGIN_MODE"
        },
        "license": {
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "devsdk.net/api/dfd",
    "basePath": "/",
    "paths": {
...
        "/v1/lol/history/updater": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find game after the timestamp and store to DB\n**Application Token** is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/lol/history"
                ],
                "summary": "Update lol history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application access token",
                        "name": "X-Dfd-App-Auth",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/lol/history/{id}": {
            "get": {
                "description": "Find game after the timestamp and store to DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/lol/history"
                ],
                "summary": "Update lol history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "game": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/lol/historys": {
            "get": {
                "description": "Find game after the timestamp and store to DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/lol/history"
                ],
                "summary": "Update lol history",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "games": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/docmodels.ResponseLoLHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.ResponseBadRequest": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Bad request message"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "docmodels.ResponseInternalServerError": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "docmodels.ResponseLoLHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "6006d3cc95f8c8e32d660c04"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1610811479544
                },
                "win": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "docmodels.ResponseNotFound": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Not found message"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "docmodels.ResponseSuccess": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "docmodels.ResponseUnauthorized": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Request Unauthorized message"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "token_expired": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}